feat: variables and clean up
