+++ src/routes/(console)/project-[project]/sites/create-site/+layout.ts
@@ -1,9 +1,12 @@
 import { sdk } from '$lib/stores/sdk';
 
 export const load = async () => {
-    const installations = await sdk.forProject.vcs.listInstallations();
-
+    const [installations, frameworks] = await Promise.all([
+        sdk.forProject.vcs.listInstallations(),
+        sdk.forProject.sites.listFrameworks()
+    ]);
     return {
-        installations
+        installations,
+        frameworks
     };
 };
